Reorganised project Structure
Fixes #007
