+++ .gitignore
@@ -1,5 +1,7 @@
 # Default ignored files
 /shelf/
-/workspace.xml
+/Week 1 Basics/.idea/workspace.xml
 # Environment-dependent path to Maven home directory
 /mavenHomeManager.xml
+**/.DS_Store
+.idea
