+++ Week 1 Basics/.idea/encodings.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="Encoding">
+    <file url="file://$PROJECT_DIR$/basics/src/main/java" charset="UTF-8" />
+  </component>
+</project>
