+++ Week 1 Basics/basics/src/main/java/com/springwithTheo/week1/basics/config/AppConfig.java
@@ -19,8 +19,9 @@ public class AppConfig {
         return new VegPizza();
     }
 
-    @Bean @Lazy
-    public  Pizza meatPizza() {
+    @Bean
+    @Lazy
+    public Pizza meatPizza() {
         return new MeatPizza();
     }
