+++ Week 1 Basics/basics/src/main/java/com/springwithTheo/week1/basics/lazy/EagerLoader.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Component;
 
 @Component
 public class EagerLoader {
-    public EagerLoader(){
+    public EagerLoader() {
         System.out.println("EagerLoader is initialized!");
     }
 }
