+++ Week 1 Basics/basics/src/main/resources/application.properties
